updating ImageDataet to use our downlaod_data function with fixes for huggingface issues
